moved test and fixed issue with auto assignment of src_indices to nonflat distrib inputs
